requirements.txt
mysql-connector
mysql-connector-python
requests
qrcode